requirements.txt
astroid==2.4.2
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.10.0
mccabe==0.6.1
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMySQL==1.0.2
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
